backend/requirements.txt
Django==2.0.2
django-ckeditor==5.4.0
django-debug-toolbar==1.9.1
django-js-asset==1.0.0
easy-thumbnails==2.5
Pillow==5.0.0
pytz==2018.3
singlemodeladmin==0.8
sqlparse==0.2.4
